extract virtual position to new file
