+++ packages/core/src/abap/nodes/statement_node.ts
@@ -1,4 +1,5 @@
-import {Position, VirtualPosition} from "../../position";
+import {Position} from "../../position";
+import { VirtualPosition } from "../../virtual_position";
 import {AbstractNode} from "./_abstract_node";
 import {INode} from "./_inode";
 import {TokenNode} from "./token_node";
