+++ packages/core/src/index.ts
@@ -30,7 +30,8 @@ import * as Tokens from "./abap/1_lexer/tokens";
 import {IConfig, IDependency, IRenameSettings, IConfiguration} from "./_config";
 import {IRegistry} from "./_iregistry";
 import {IFile} from "./files/_ifile";
-import {Position, VirtualPosition} from "./position";
+import {Position} from "./position";
+import { VirtualPosition } from "./virtual_position";
 import {AbstractFile} from "./files/_abstract_file";
 import {PrettyPrinter} from "./pretty_printer/pretty_printer";
 import {ISpaghettiScope, ISpaghettiScopeNode} from "./abap/5_syntax/_spaghetti_scope";
