+++ packages/core/src/objects/rename/renamer_helper.ts
@@ -5,7 +5,7 @@ import {Identifier} from "../../abap/4_file_information/_identifier";
 import {SyntaxLogic} from "../../abap/5_syntax/syntax";
 import {ScopeType} from "../../abap/5_syntax/_scope_type";
 import {ISpaghettiScopeNode} from "../../abap/5_syntax/_spaghetti_scope";
-import {VirtualPosition} from "../../position";
+import { VirtualPosition } from "../../virtual_position";
 import {IRegistry} from "../../_iregistry";
 import {ABAPObject} from "../_abap_object";
 import {AbstractObject} from "../_abstract_object";
