+++ packages/core/src/pretty_printer/indent.ts
@@ -3,7 +3,7 @@ import * as Expressions from "../abap/2_statements/expressions";
 import {MacroContent, Comment, Empty} from "../abap/2_statements/statements/_statement";
 import {StatementNode} from "../abap/nodes/statement_node";
 import {IIndentationOptions} from "./indentation_options";
-import {VirtualPosition} from "../position";
+import { VirtualPosition } from "../virtual_position";
 import {ABAPFile} from "../abap/abap_file";
 
 // todo, will break if there is multiple statements per line?
