+++ packages/core/src/rules/downport.ts
@@ -8,7 +8,8 @@ import {IRule, IRuleMetadata, RuleTag} from "./_irule";
 import {Unknown} from "../abap/2_statements/statements/_statement";
 import {ExpressionNode, StatementNode, TokenNode} from "../abap/nodes";
 import {IEdit, EditHelper} from "../edit_helper";
-import {Position, VirtualPosition} from "../position";
+import {Position} from "../position";
+import { VirtualPosition } from "../virtual_position";
 import {ABAPFile} from "../abap/abap_file";
 import {IRegistry} from "../_iregistry";
 import {IObject} from "../objects/_iobject";
