+++ packages/core/src/rules/indentation.ts
@@ -9,7 +9,8 @@ import * as Statements from "../abap/2_statements/statements";
 import * as Expressions from "../abap/2_statements/expressions";
 import {IRuleMetadata, RuleTag} from "./_irule";
 import {DDIC} from "../ddic";
-import {Position, VirtualPosition} from "../position";
+import {Position} from "../position";
+import { VirtualPosition } from "../virtual_position";
 import {EditHelper} from "../edit_helper";
 import {ABAPFile} from "../abap/abap_file";
